#53 extract colormanager into independent file
