+++ Assets/Scripts/Book/ColorManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08eec99ad7f464f439820b843e03607f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
